Working example with base project dependency
