+++ _WORKSPACE.bzlmod
@@ -1,0 +1,7 @@
+# See:
+# - https://www.youtube.com/watch?v=2Nn71RV_yhI
+# - https://bazel.build/external/overview#blzmod
+# - https://docs.aspect.build/rules/aspect_bazel_lib/
+# - https://github.com/aspect-build/bazel-examples/blob/main/oci_python_image/MODULE.bazel
+
+# NOTE: Currently not used since it overrides WORKSPACE.bazel
