+++ projects/base_py_fastapi_app/app/main.py
@@ -1,0 +1,46 @@
+import click
+import logging
+import random
+
+from fastapi import FastAPI
+from fastapi.responses import JSONResponse
+
+# from libs.devops.models.devops import DevOps, InfrastructureEngineer, DeveloperExperienceEngineer, DataEngineer, MachineLearningEngineer, WebEngineer, ReliabilityEngineer, PlatformEngineer, PlatformOrganization
+
+# TODO: Wrap this project into a class maybe?
+# class DevOps_v2:
+#     def __init__(self, name: str) -> None:
+#         self.name = name
+
+#     def __str__(self) -> str:
+#         raise NotImplementedError
+
+#     def speak(self) -> None:
+#         raise NotImplementedError
+
+#     def responsibility(self) -> None:
+#         raise NotImplementedError
+
+# See: https://docs.python.org/3/library/logging.html
+logging.basicConfig(level=logging.INFO, format='%(asctime)s - %(levelname)s : %(module)s %(message)s')
+logging.getLogger(__name__)
+
+app = FastAPI()
+
+
+@app.on_event("startup")
+async def startup_event():
+    logging.info(f"=== [Starting FastAPI Server...] ===")
+
+@app.on_event("shutdown")
+async def shutdown_event():
+    logging.info(f"=== [Stopping FastAPI Server...] ===")
+
+
+@app.get("/")
+async def root():
+    return JSONResponse("I am alive!!!")
+
+@app.get("/status")
+async def read_root():
+    return {"status": "UP", "version": "0.1.2"}
